requirements.txt
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
boto==2.34.0
diff-match-patch==20110725.1
dj-database-url==0.3.0
dj-static==0.0.5
django-autoslug==1.7.2
django-registration==1.0
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.2
static==1.0.2
tablib==0.10.0
virtualenv==1.11.6
wsgiref==0.1.2
